+++ meeshkan/__init__.py
@@ -1,0 +1,5 @@
+HttpExchangeStream = AsyncIterable[HttpExchange]
+BuildResultStream = AsyncIterable[BuildResult]
+
+Source = Callable[[AbstractEventLoop], HttpExchangeStream]
+Sink = Callable[[BuildResultStream], Awaitable[None]]
